refactor(sdk/go): dedup encodeConnectEnvelope and fix data-plane bugs
Merge duplicate encodeConnectEnvelope into a shared connect.go; add pty.go reusing it. Fix envd RPC port (JupyterPort->EnvdPort 49983), exit-0 recovery from status/exited, and CreateSnapshot empty-body 422. Add unit tests.
